+++ CHANGELOG.md
@@ -14,8 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SDK: Ability for rovingFocusContainer to change focused item for lazy loading containers
 - CI, INTERNAL: Add and enforce ktfmt formatting
 - SDK,UI: Add delete avatar functionality
+- SDK: Sending audio files now sets `info.duration` if possible
 
 ### Changed
+
+- SDK: (**breaking**) moved `FileKitFileDescriptor` from `de.connect2x.trixnity.messenger.compose.view.files` to
+  `de.connect2x.trixnity.messenger.util`
 
 ### Deprecated